Whole-app end-user probe: four ways a surface said one thing and did another (#36)
Every probe so far took one milestone. This one drove the whole app the way a new
user meets it — the six guide pages command by command, `loom infer` → `validate` →
`plan` → `apply` → `rollback` against a real SQLite/filesystem warehouse, the write
plane over MCP under governance, and the attested-identity plane against a real RSA
JWKS issuer over HTTP.

Most of it came back clean, including two planes no probe had driven end to end:
`run_<action>` over stdio with `edit_log: required` (a deleted ticket's vector pruned
before its row, a refused run still recorded, `dryRun` recording nothing), and
`mcp.auth` + `when:` over HTTP — gold-desk token sees 2 rows, staff token sees 4, a
token *missing* the claim leaves the guard undecided so the policy applies, and bad
audience / bad issuer / expired / absent / garbage all get 401 with a proper
`WWW-Authenticate`. A gold-desk caller writing to a row their own policy hides gets
`object_not_found`, and `principal` lands in `_loom_meta.edits` beside `actor`.

Four things it found, each fixed with the test that would have caught it.

**`int -> double` is not an Iceberg promotion, and three commands disagreed about it.**
`types._PROMOTIONS` was written as though it were Iceberg's ALTER-time set and was not:
Iceberg promotes `int -> long` and `float -> double` and nothing else among the types a
spec can name. Against a real catalog `loom validate --physical` said ok, `loom plan`
said physical-safe, and `loom apply` got `Cannot change column type` out of pyiceberg
part way through the run — appending a `partial` row to `_loom_meta.applied` for a run
that committed nothing. The two questions are now two functions: `promotable` is the
read question physical validation asks, `iceberg_alterable` is the DDL question the
planner asks. Anything outside Iceberg's set is breaking, refused before the catalog is
touched. `docs/guide/migrations.md` used this exact change as its worked example of
physical-safe. Every planner test built columns by hand against a fake catalog and the
pyiceberg-backed apply tests only ever ran `int -> long`, so both halves agreed with
each other without either asking Iceberg; the regression test runs the refused plan
against a real one.

**`loom.yaml` inside the ontology directory was fatal.** `find_config` looks there
*first* and documents the order; `load_dir` globbed every `*.yaml` under the same
directory and demanded each declare a spec kind, so the config file's own preferred
location made every command exit 1 with an error about spec grammar naming a file that
is not a spec. Only the one at the ontology root is skipped — a nested `loom.yaml` is
config `find_config` would never read, so it keeps erroring.

**`loom infer data.csv` never reached the refusal built for it.** CSV and JSON are
refused by name with the reason, and that refusal fired only if you typed `--format
csv`. `--format` still derives no reader from the extension; it now decides which
refusal you get to read instead of pyarrow's "Parquet magic bytes not found in footer".

**A quarantine described itself as a refusal.** `loom ingest --reject-to` applies —
one row written, two set aside, exit 0 — and said otherwise at every step: the preview
above the `y/N` marked the set-aside rows `!` and printed `nothing was written.`, and
the summary printed one `error:` line per row, indistinguishable from a refused load.
`QUARANTINABLE` moved into `ingest.result` so the runtime and the CLI use one set;
quarantined rows render `·` and report `rejected:`.

Plus the doc drift the probe walked into: the quickstart's `1091 tests` (1177), its
`10 MCP tools` (14), and an install line that omits the `embed` extra on a page that
gets as far as the ranked plane.

+++ tests/test_apply.py
@@ -195,7 +195,10 @@ def test_every_edit_to_a_table_is_applied_in_a_single_call(ontology, snapshot):
     Iceberg transaction, so two calls would be two commits and a window where half of a column's
     migration had landed. `lifetime_value` here needs both halves: a widening and a loosening."""
     live = dict(CUSTOMERS)
-    live["lifetime_value"] = Column("lifetime_value", "int", required=True, field_id=4)
+    # `float`, because Iceberg's promotion set is `int -> long` and `float -> double` — an `int`
+    # column under this `double` property reads fine and cannot be *altered* into one, which is a
+    # breaking plan rather than the two-edit alter this test is about.
+    live["lifetime_value"] = Column("lifetime_value", "float", required=True, field_id=4)
     catalog = FakeWritableCatalog(tables={"crm.customers": live, "sales.orders": ORDERS})
 
     result = apply_plan(_plan(ontology, catalog), {catalog.name: catalog}, snapshot)
@@ -224,7 +227,7 @@ def test_a_missing_optional_column_is_added(ontology, snapshot):
 
 def test_a_promotion_is_applied_as_a_type_update(ontology, snapshot):
     narrowed = dict(CUSTOMERS)
-    narrowed["lifetime_value"] = Column("lifetime_value", "int", required=False, field_id=4)
+    narrowed["lifetime_value"] = Column("lifetime_value", "float", required=False, field_id=4)
     catalog = FakeWritableCatalog(tables={"crm.customers": narrowed, "sales.orders": ORDERS})
 
     apply_plan(_plan(ontology, catalog), {catalog.name: catalog}, snapshot)
